requirements.txt
asgiref==3.7.2
Django==3.2.4
djangorestframework==3.14.0
flake8==6.1.0
mccabe==0.7.0
pycodestyle==2.11.0
pyflakes==3.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
